Change stat_total_boat_production to float

+++ app/database/migrations/2019_07_28_175059_add_dominion_statistics_fields.php
@@ -25,7 +25,7 @@ class AddDominionStatisticsFields extends Migration
             $table->unsignedInteger('stat_total_ore_production')->after('stat_total_mana_production')->default(0);
             $table->unsignedInteger('stat_total_gem_production')->after('stat_total_ore_production')->default(0);
             $table->unsignedInteger('stat_total_tech_production')->after('stat_total_gem_production')->default(0);
-            $table->unsignedInteger('stat_total_boat_production')->after('stat_total_tech_production')->default(0);
+            $table->float('stat_total_boat_production')->after('stat_total_tech_production')->default(0);
         });
     }
